+++ Sources/KeyboardKit/Input/AlphabeticKeyboardInputSet.swift
@@ -1,0 +1,14 @@
+//
+//  AlphabeticKeyboardInputSet.swift
+//  KeyboardKit
+//
+//  Created by Daniel Saidi on 2021-01-30.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+/**
+ This input set can be used in alphabetic keyboards.
+ */
+public class AlphabeticKeyboardInputSet: KeyboardInputSet {}