+++ Sources/KeyboardKit/Input/KeyboardInputSet.swift
@@ -9,17 +9,13 @@
 import Foundation
 
 /**
- A "keyboard input set" represents the text input parts of a
- system keyboard, i.e. the lighter input keys.
+ A keyboard input set represents the input parts of a system
+ keyboard, the center lighter input keys.
  
- Some devices may add additional inputs to an input set. For
+ Some devices add additional inputs to system keyboards. For
  instance, iPad devices adds punctionation and characters to
  the bottom rows. These inputs are not considered to be part
- of the input set, but rater the full keyboard layout.
- 
- To implement a system keyboard, you need a `KeyboardLayout`
- instead of just an input set, since a layout also specifies
- the actions that surround the input set characters.
+ of the input set, but rater the keyboard layout.
  */
 public class KeyboardInputSet: Equatable {
     
@@ -35,18 +31,3 @@ public class KeyboardInputSet: Equatable {
         lhs.inputRows == rhs.inputRows
     }
 }
-
-/**
- This input set should only be used in alphabetic keyboards.
- */
-public class AlphabeticKeyboardInputSet: KeyboardInputSet {}
-
-/**
- This input set should only be used in numeric keyboards.
- */
-public class NumericKeyboardInputSet: KeyboardInputSet {}
-
-/**
- This input set should only be used in symbolic keyboards.
- */
-public class SymbolicKeyboardInputSet: KeyboardInputSet {}