+++ Sources/KeyboardKit/Input/Providers/GermanKeyboardInputSetProvider.swift
@@ -1,0 +1,33 @@
+//
+//  GermanKeyboardInputSetProvider.swift
+//  KeyboardKit
+//
+//  Created by Daniel Saidi on 2020-12-01.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+/**
+ This class provides German keyboard input sets.
+ */
+public class GermanKeyboardInputSetProvider: KeyboardInputSetProvider {
+    
+    public init() {}
+    
+    public func alphabeticInputSet(for context: KeyboardContext) -> AlphabeticKeyboardInputSet {
+        AlphabeticKeyboardInputSet(inputRows: [
+            ["q", "w", "e", "r", "t", "z", "u", "i", "o", "p", "ü"],
+            ["a", "s", "d", "f", "g", "h", "j", "k", "l", "ö", "ä"],
+            ["y", "x", "c", "v", "b", "n", "m"]
+        ])
+    }
+    
+    public func numericInputSet(for context: KeyboardContext) -> NumericKeyboardInputSet {
+        .standard(currency: "€")
+    }
+    
+    public func symbolicInputSet(for context: KeyboardContext) -> SymbolicKeyboardInputSet {
+        .standard(currencies: ["$", "£", "¥"])
+    }
+}