+++ Tests/KeyboardKitTests/Input/AlphabeticKeyboardInputSetTests.swift
@@ -1,0 +1,16 @@
+//
+//  AlphabeticKeyboardInputSetTests.swift
+//  KeyboardKit
+//
+//  Created by Daniel Saidi on 2021-01-30.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
+//
+
+import Quick
+import Nimble
+import KeyboardKit
+
+class AlphabeticKeyboardInputSetTests: QuickSpec {
+    
+    override func spec() {}
+}