+++ Tests/KeyboardKitTests/Input/StandardKeyboardInputSetProviderTests.swift
@@ -18,9 +18,19 @@ class StandardKeyboardInputSetProviderTests: QuickSpec {
         var provider: KeyboardInputSetProvider!
         var context: MockKeyboardContext!
         
+        var english: KeyboardInputSetProvider!
+        var german: KeyboardInputSetProvider!
+        var italian: KeyboardInputSetProvider!
+        var swedish: KeyboardInputSetProvider!
+        
         beforeEach {
             provider = StandardKeyboardInputSetProvider()
             context = MockKeyboardContext()
+            
+            english = EnglishKeyboardInputSetProvider()
+            german = GermanKeyboardInputSetProvider()
+            italian = ItalianKeyboardInputSetProvider()
+            swedish = SwedishKeyboardInputSetProvider()
         }
         
         describe("standard keyboard input set provider") {
@@ -42,44 +52,44 @@ class StandardKeyboardInputSetProviderTests: QuickSpec {
             
             it("supports English") {
                 let locale = "en"
-                expect(alphabetic(for: locale)).to(equal(.alphabetic_en))
-                expect(numeric(for: locale)).to(equal(.numeric_en))
-                expect(symbolic(for: locale)).to(equal(.symbolic_en))
+                expect(alphabetic(for: locale)).to(equal(english.alphabeticInputSet(for: context)))
+                expect(numeric(for: locale)).to(equal(english.numericInputSet(for: context)))
+                expect(symbolic(for: locale)).to(equal(english.symbolicInputSet(for: context)))
             }
             
             it("supports German") {
                 let locale = "de"
-                expect(alphabetic(for: locale)).to(equal(.alphabetic_de))
-                expect(numeric(for: locale)).to(equal(.numeric_de))
-                expect(symbolic(for: locale)).to(equal(.symbolic_de))
+                expect(alphabetic(for: locale)).to(equal(german.alphabeticInputSet(for: context)))
+                expect(numeric(for: locale)).to(equal(german.numericInputSet(for: context)))
+                expect(symbolic(for: locale)).to(equal(german.symbolicInputSet(for: context)))
             }
             
             it("supports Italian") {
                 let locale = "it"
-                expect(alphabetic(for: locale)).to(equal(.alphabetic_it))
-                expect(numeric(for: locale)).to(equal(.numeric_it))
-                expect(symbolic(for: locale)).to(equal(.symbolic_it))
+                expect(alphabetic(for: locale)).to(equal(italian.alphabeticInputSet(for: context)))
+                expect(numeric(for: locale)).to(equal(italian.numericInputSet(for: context)))
+                expect(symbolic(for: locale)).to(equal(italian.symbolicInputSet(for: context)))
             }
             
             it("supports Swedish") {
                 let locale = "sv"
-                expect(alphabetic(for: locale)).to(equal(.alphabetic_sv))
-                expect(numeric(for: locale)).to(equal(.numeric_sv))
-                expect(symbolic(for: locale)).to(equal(.symbolic_sv))
+                expect(alphabetic(for: locale)).to(equal(swedish.alphabeticInputSet(for: context)))
+                expect(numeric(for: locale)).to(equal(swedish.numericInputSet(for: context)))
+                expect(symbolic(for: locale)).to(equal(swedish.symbolicInputSet(for: context)))
             }
             
             it("has fallback support for specific locale") {
                 let locale = "en-US"
-                expect(alphabetic(for: locale)).to(equal(.alphabetic_en))
-                expect(numeric(for: locale)).to(equal(.numeric_en))
-                expect(symbolic(for: locale)).to(equal(.symbolic_en))
+                expect(alphabetic(for: locale)).to(equal(english.alphabeticInputSet(for: context)))
+                expect(numeric(for: locale)).to(equal(english.numericInputSet(for: context)))
+                expect(symbolic(for: locale)).to(equal(english.symbolicInputSet(for: context)))
             }
             
             it("has fallback support for non-existing locale") {
                 let locale = "abc"
-                expect(alphabetic(for: locale)).to(equal(.alphabetic_en))
-                expect(numeric(for: locale)).to(equal(.numeric_en))
-                expect(symbolic(for: locale)).to(equal(.symbolic_en))
+                expect(alphabetic(for: locale)).to(equal(english.alphabeticInputSet(for: context)))
+                expect(numeric(for: locale)).to(equal(english.numericInputSet(for: context)))
+                expect(symbolic(for: locale)).to(equal(english.symbolicInputSet(for: context)))
             }
         }
     }